merging latest tensorflow

+++ tensorflow/lite/experimental/micro/micro_interpreter.cc
@@ -14,14 +14,14 @@ limitations under the License.
 ==============================================================================*/
 #include "tensorflow/lite/experimental/micro/micro_interpreter.h"
 
-#include "tensorflow/lite/c/c_api_internal.h"
+#include "tensorflow/lite/c/common.h"
 #include "tensorflow/lite/core/api/flatbuffer_conversions.h"
 #include "tensorflow/lite/experimental/micro/compatibility.h"
 #include "tensorflow/lite/experimental/micro/micro_optional_debug_tools.h"
 
 namespace tflite {
 namespace {
-const int kStackDataAllocatorSize = 128;
+const size_t kStackDataAllocatorSize = 128;
 class StackDataAllocator : public BuiltinDataAllocator {
  public:
   void* Allocate(size_t size) override {
@@ -92,7 +92,7 @@ MicroInterpreter::MicroInterpreter(const Model* model,
   // NOTE: This requires that the flatbuffer is held in memory which can be
   // modified by this process.
   if (!FLATBUFFERS_LITTLEENDIAN) {
-    for (int t = 0; t < tensors_size(); ++t) {
+    for (size_t t = 0; t < tensors_size(); ++t) {
       TfLiteTensor* thisTensor = &context_.tensors[t];
       if (thisTensor->allocation_type == kTfLiteMmapRo)
         CorrectTensorEndianness(thisTensor);
